refactor: move Slack test files to correct directory
